+++ Changelog.md
@@ -2,6 +2,7 @@
 ### 7.2.0.0 (not yet released)
 
 - Next/Previous uses only known media file types
+  as defined in the settings
 - Larger list of known media file types
 - Bit depth parameter is hidden for DTS